More correctly doing HyperLogLog math
Updates the HyperLogLog math to correct an issue with generating estimates
that were ~50% of the expected value. Unclear exactly why this is the case,
but this fixes the math.

+++ libhyperloglog.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-typedef uint64_t (*hash_func)(const void *key, int len, uint32_t seed);
+typedef uint64_t (*hash_func)(const void *key, int len, uint64_t seed);
 typedef struct hyperloglog count_t;
 
 // Creates a new filter
